Explicit proxy value should take precedence.
For those of us in the corporate world we may be behind several proxies
if we're accessing services that other departments (or even vendors) are
providing. There is a very real use case for having an environment proxy
set for general HTTP service calls and then explicitly, for let's say
OpenStack, needing to override that.

I ran into this problem using Fog while attempting to connect to an
internal OpenStack cluster of machines that is in a separate segment of
our network.