change travis tests rbx-2.1.1 -> rbx-2
This removes testing of the slightly older 2.1.1 release which is currently failing in TravisCI. This is due to missing bundler, which is officially packages as of 2.2.5 but works with travis starting from 2.2.0 (it uses rvm context).

With this tests move to the most current rubinius build, which is 2.2.6 at the moment.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>