stream first, sleep later for streaming tests
When sleeping first and streaming second, in case of Content-Length headers there results will come back and the test will end while the server still has to go for another round of sleep 1s. This breaks following tests. To get around this issue, always sleep 1s first, then stream a piece of the result. The tests only need minimal adjustments to make it work and the purpose is still covered.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>