Adds defaults option to keep cookies through redirects
To pass cookies set by the initial location to the redirection
location, set this defaults option:

    Excon.defaults[:redirect_with_cookies] = true

I encountered a use case where a response set a cookie and supplied
a redirect location. Excon would redirect correctly, but the new
location required the cookie value set by the initial response.

It seems to me that this would be the default functionality, but I
instead chose to make it opt-in since it's new functionality.