Don't override Host when talking to UNIX socket
As per the RFC, sending Host="" is the correct behavior when speaking to
a URI that has no host (e.g. a UNIX socket):

> If the requested URI does not include an Internet host name for the
> service being requested, then the Host header field MUST be given with
> an empty value

Unfortunately, some web servers treat that the same as an entirely
omitted Host header, which results in a 400 response. To work around
this, one can fabricate and send any valid Host to which the server will
respond.

For Excon to support this use-case, it needs to respect an explicitly
set Host header, even when talking to a UNIX socket.