Raise on warnings in tests
Currently Excon tests trigger a `report_warning` in several places that
indicate bugs in the code (missing items in the list of valid parameter keys).
The tests don't display these warnings so it's difficult to catch this kind of
regression. This PR makes it such that tests fail on warnings and updates the
code to avoid warnings in tests. This will also facilitate testing the changes
in #289 since that will add new warnings and validation behavior.