[fix] Read server response during EPIPE
When a `request_block` is used to send data, an error on the server side
only gets reported as a `EPIPE`. excon doesn't read anything sent back
from the server, which makes it hard to debug what went wrong.

We now read the server response, if any, and show it in the exception.

Closes #760