tests: ensure we always run the current Ruby interpreter
In Debian, sometimes we have multiple Ruby interpreters installed at the
same time, but only one of them is the default `ruby`. For example at
the moment we are transitioning between ruby2.7 (the current default)
and ruby3.0. When running the tests against ruby3.0, we want to ensure
that everything else that starts a new Ruby interpreter starts ruby3.0,
not the default `ruby`.