Use system OpenSSL cert file and dir if available
For users that manage their own SSL certs, it is a surprise that excon
ships its own root certificates. As a result, using an old version of
excon can cause SSL failures with third-party services, such as S3
(https://github.com/excon/excon/issues/473).

We now fall back to the OpenSSL defaults if available to avoid depending
on the shipped `cacert.pem`.