Automatic Fork Safety
Forking servers are a popular deployment method in Ruby.
One big footgun is that uppon fork, all file descriptors
are inherited, so if you create some FDs before fork and hold on them
you may end up with the same connection being used across many process
creating havoc.

As such it's a good idea for libraries that do keep persistent
connections do check the PID before re-using them.

The alternative is for libraries to provide some kind of callback that
the application has to call before fork, but that's easilly missed
by users and the PID check is very cheap.